# ext/bcmath: Use SIMD for trailing zero counts during conversion