@SakiTakamachi 2024-05-07
@nielsdos 

Does this patch mean that Benchmark 2 is a bit slower in your environment?

Could it be that the speedup in my measurements with this patch is due to the use of UNEXPECTED and the removal of unnecessary code, and that SIMD has a negative effect on patch 2?
(I am concerned that, as you said before, measurements may be faster or slower immediately after compilation.)