@SakiTakamachi 2024-05-08
> Do we think trailing zeros is common?

Trailing zeros are probably quite common given the use cases for BCMath, but 16 decimal digits is probably quite rare.

I've opened a PR on this as it improved performance in all cases in my environment, but if not, I wouldn't be picky about using SIMD here.