@nielsdos 2024-05-08
These are the results I'm getting:

```
Benchmark 1: ./sapi/cli/php 1.php
  Time (mean ± σ):     468.3 ms ±   9.7 ms    [User: 463.6 ms, System: 1.9 ms]
  Range (min … max):   457.7 ms … 486.1 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old 1.php
  Time (mean ± σ):     450.9 ms ±   3.6 ms    [User: 448.1 ms, System: 2.5 ms]
  Range (min … max):   446.3 ms … 457.0 ms    10 runs
 
Summary
  ./sapi/cli/php_old 1.php ran
    1.04 ± 0.02 times faster than ./sapi/cli/php 1.php
```

```
Benchmark 1: ./sapi/cli/php 2.php
  Time (mean ± σ):     535.1 ms ±  18.0 ms    [User: 531.7 ms, System: 2.8 ms]
  Range (min … max):   517.8 ms … 578.0 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old 2.php
  Time (mean ± σ):     527.9 ms ±  12.1 ms    [User: 525.9 ms, System: 1.5 ms]
  Range (min … max):   517.1 ms … 552.8 ms    10 runs
 
Summary
  ./sapi/cli/php_old 2.php ran
    1.01 ± 0.04 times faster than ./sapi/cli/php 2.php
```

```
Benchmark 1: ./sapi/cli/php 3.php
  Time (mean ± σ):     496.5 ms ±   8.1 ms    [User: 493.8 ms, System: 2.2 ms]
  Range (min … max):   490.5 ms … 515.1 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old 3.php
  Time (mean ± σ):     613.2 ms ±  19.1 ms    [User: 610.5 ms, System: 2.2 ms]
  Range (min … max):   602.9 ms … 666.7 ms    10 runs

Summary
  ./sapi/cli/php 3.php ran
    1.24 ± 0.04 times faster than ./sapi/cli/php_old 3.php
```