@SakiTakamachi 2024-05-09
@nielsdos 
I also merged to this the latest master and compared it with EC2.

master:
```
hyperfine "php 1.php" --warmup 10
Time (mean ± σ):     654.7 ms ±   3.0 ms    [User: 650.5 ms, System: 2.6 ms]
Range (min … max):   650.1 ms … 659.2 ms    10 runs

hyperfine "php 2.php" --warmup 10
Time (mean ± σ):     769.4 ms ±   5.4 ms    [User: 765.8 ms, System: 2.0 ms]
Range (min … max):   762.6 ms … 781.8 ms    10 runs

hyperfine "php 3.php" --warmup 10
Time (mean ± σ):     910.3 ms ±  13.3 ms    [User: 905.6 ms, System: 2.8 ms]
Range (min … max):   896.8 ms … 934.6 ms    10 runs

php old.php // my old bench
1.6298861503601
1.9048039913177
2.2188358306885
```

this branch:
```
hyperfine "php 1.php" --warmup 10
Time (mean ± σ):     643.6 ms ±   6.5 ms    [User: 638.7 ms, System: 3.2 ms]
Range (min … max):   637.0 ms … 656.7 ms    10 runs

hyperfine "php 2.php" --warmup 10
Time (mean ± σ):     749.7 ms ±   6.4 ms    [User: 745.4 ms, System: 2.4 ms]
Range (min … max):   742.3 ms … 766.6 ms    10 runs

hyperfine "php 3.php" --warmup 10
Time (mean ± σ):     684.7 ms ±  10.7 ms    [User: 680.8 ms, System: 2.5 ms]
Range (min … max):   673.4 ms … 707.8 ms    10 runs

php old.php // my old bench
1.5792031288147
1.8460278511047
1.6792199611664
```