@SakiTakamachi 2024-05-09
I removed the unnecessary cast and changed the variable name slightly. If the variable names are okay, merge this.